Fix logic issues and unify IndexedDB handling
