+++ CODEX_UPDATE_M29.md
@@ -1,0 +1,21 @@
+# Code Updates for Milestone M29
+
+The following logic adjustments and cleanups were applied based on the previous audit in `NOT_LOGICAL.md`.
+
+## Shared IndexedDB Helper
+- Created `frontend/src/utils/indexedDb.ts` which exports `openCredentialDB`.
+- `credentialVault.ts` and `encryption.ts` now import and use this helper instead of having separate implementations.
+
+## Comment and Import Cleanup
+- Clarified the comment for `PBKDF2_ITERATIONS` in `encryption.ts`.
+- Removed unused `querystring` import from `backend/ipc/oauthServer.ts`.
+- Removed unsupported `userDataDir` option from Playwright context creation.
+
+## Documentation Updated
+- Updated `NOT_LOGICAL.md` to reflect resolved issues.
+
+## Areas for Cursor Review
+1. Ensure `openCredentialDB` correctly handles upgrade scenarios when future stores or versions are added.
+2. Review Playwright automation for any other unsupported parameters or missing error handling.
+3. Verify encryption key storage and retrieval still work after switching to the shared IndexedDB helper.
+
